Modularize base query
